# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vector DB",
    "version": "3.0.0"
  },
  "paths": {
...
    "/v1/collections/{collection_name}/upsert": {
      "post": {
        "tags": [
          "vectors"
        ],
        "summary": "Upsert Vector In Collection",
        "operationId": "upsert_vector_in_collection_v1_collections__collection_name__upsert_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Item"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collections/{collection_name}/bulk_upsert": {
      "post": {
        "tags": [
          "vectors"
        ],
        "summary": "Bulk Upsert In Collection",
        "operationId": "bulk_upsert_in_collection_v1_collections__collection_name__bulk_upsert_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpsertRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collections/{collection_name}/delete/{external_id}": {
      "delete": {
        "tags": [
          "vectors"
        ],
        "summary": "Delete Vector In Collection",
        "operationId": "delete_vector_in_collection_v1_collections__collection_name__delete__external_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          },
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/collections/{collection_name}/delete_batch": {
      "post": {
        "tags": [
          "vectors"
        ],
        "summary": "Batch Delete In Collection",
        "operationId": "batch_delete_in_collection_v1_collections__collection_name__delete_batch_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDeleteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/upsert": {
      "post": {
        "tags": [
          "vectors"
        ],
        "summary": "Upsert Vector Legacy",
        "operationId": "upsert_vector_legacy_v1_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Item"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/bulk_upsert": {
      "post": {
        "tags": [
          "vectors"
        ],
        "summary": "Bulk Upsert Legacy",
        "operationId": "bulk_upsert_legacy_v1_bulk_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/delete/{external_id}": {
      "delete": {
        "tags": [
          "vectors"
        ],
        "summary": "Delete Vector Legacy",
        "operationId": "delete_vector_legacy_v1_delete__external_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "External Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchDeleteRequest": {
        "properties": {
          "external_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "External Ids"
          }
        },
        "type": "object",
        "required": [
          "external_ids"
        ],
        "title": "BatchDeleteRequest"
      },
      "BulkUpsertRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UpsertRequest"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BulkUpsertRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UpsertRequest": {
        "properties": {
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "vector": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Vector"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "vector"
        ],
        "title": "UpsertRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
}